docs: add an alternative package if original fails
The old instructions mentioned adafruit-python**3**-shell as a dependency but when I tried to install the package I found out that the package didn't exist ( or deprecated I think ) and I was able to use adafruit-python-shell instead. So, I thought I might add the package to the instructions in case anyone else runs into the same issue as me.